Feat : 임시뷰가 보이면서 드래그하면 사각형 또는 사진이 이동하는 기능 구현  임시뷰를 만들면서 Rectangle객체를 굳이 클론할 필요가 없어서 클론 기능 삭제 지저분한 코드들 리팩토링 태블릿이 아닌 일반 기기에서도 작동은 할 수 있도록 레이아웃 보완
