+++ app/src/main/java/com/codesquard/kotlin_drawingapp/NormalRectangle.kt
@@ -7,14 +7,4 @@ class NormalRectangle : Rectangle {
     override lateinit var color: Array<Int>
     override var isSelected: Boolean = false
     override var alphaValue: Int = 0
-
-    override fun clone(): NormalRectangle {
-        val clonedRect = NormalRectangle()
-        clonedRect.apply {
-            this.size = this@NormalRectangle.size
-            this.color = this@NormalRectangle.color
-            this.point = this@NormalRectangle.point
-        }
-        return clonedRect
-    }
 }
