+++ app/src/main/java/com/codesquard/kotlin_drawingapp/Rectangle.kt
@@ -32,6 +32,4 @@ interface Rectangle {
         this.isSelected = boolean
     }
 
-    fun clone(): Rectangle
-
 }
