+++ app/src/main/res/layout/activity_main.xml
@@ -16,16 +16,34 @@
     android:drawableTop="@drawable/ic_crop_square_black_24dp"
     android:text="@string/rectangle"
     app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toStartOf="@+id/photo_btn"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintHorizontal_chainStyle="packed"/>
+  <Button
+    android:id="@+id/photo_btn"
+    style="?attr/materialButtonOutlinedStyle"
+    android:layout_width="90dp"
+    android:layout_height="70dp"
+    android:backgroundTint="#E9E6FA"
+    android:drawableTop="@drawable/ic_add_photo_alternate_black_outline_24dp"
+    android:text="@string/photo"
+    app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent"/>
+    app:layout_constraintStart_toEndOf="@id/create_btn"/>
   <com.codesquard.kotlin_drawingapp.CustomView
-    android:id="@+id/customView"
+    android:id="@+id/custom_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent"
     app:layout_constraintEnd_toStartOf="@+id/second_constraintLayout"
     app:layout_constraintBottom_toTopOf="@+id/create_btn"/>
+  <com.codesquard.kotlin_drawingapp.TemporaryView
+    android:id="@+id/temporary_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent"/>
   <androidx.constraintlayout.widget.ConstraintLayout
     android:id="@+id/second_constraintLayout"
     android:layout_width="150dp"
@@ -47,10 +65,10 @@
   <Button
     android:id="@+id/background_btn"
     style="?attr/materialButtonOutlinedStyle"
-    android:layout_width="wrap_content"
+    android:layout_width="130dp"
     android:layout_height="wrap_content"
     android:layout_marginTop="10dp"
-    android:layout_marginRight="50dp"
+    android:layout_marginRight="10dp"
     app:layout_constraintEnd_toEndOf="parent"
     app:layout_constraintTop_toBottomOf="@+id/text_background"/>
   <TextView
